requirements.txt
actionlib==1.14.0
angles==1.9.13
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
docker-pycreds==0.4.0
drone-wrapper==1.4.2
dynamic-reconfigure==1.7.3
dynamixel_sdk==3.7.51
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gazebo_video_monitor_utils==0.7.0
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
image-geometry==1.16.2
importlib_resources==6.4.0
imutils==0.5.4
interactive-markers==1.12.0
Jinja2==3.1.4
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.5
laser_geometry==1.6.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mavros==1.17.0
message-filters==1.16.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
platformdirs==4.2.1
pr2_power_board==1.1.10
protobuf==4.25.3
psutil==5.9.8
py-trees==0.7.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
PyYAML==6.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
rakuda2==0.0.0
requests==2.31.0
resource_retriever==1.12.7
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-console==0.4.12
rqt-controller-manager==0.20.0
rqt-drone-teleop==1.4.2
rqt-ez-publisher==0.6.1
rqt-ground-robot-teleop==1.4.2
rqt-image-view==0.4.17
rqt-joint-trajectory-controller==0.21.2
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-multiplot==0.0.12
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_pr2_dashboard==0.4.0
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_py_trees==0.4.1
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.20
sensor-msgs==1.13.1
sentry-sdk==2.1.1
setproctitle==1.3.3
six==1.16.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
smmap==5.0.1
sympy==1.12
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
topic-tools==1.16.0
torch==2.3.0
torchvision==0.18.0
triton==2.3.0
typing_extensions==4.11.0
unique_id==1.0.6
urllib3==2.2.1
wandb==0.17.0
xacro==1.14.16
zipp==3.18.1
